Added support for cuda:N to choose GPU. Added arg for --eval-every to control frequency of validation. Added arg for --num-workers and fix the code so it will work. Added --compile-model to run torch.compile on the model.
On a fast GPU you need a lot of load workers to keep the GPU busy.
They are doing numpy transformation for data augmention on the cpu
and it is quite slow.  TODO: Use the v2 transform that allows the
transformation directly on Tensors such that they can be done on
the GPU.